# Conflicts:
#	valuation-app/build.gradle

+++ valuation-app/src/main/java/com/acuo/valuation/providers/acuo/PricingResultPersister.java
@@ -1,7 +1,7 @@
 package com.acuo.valuation.providers.acuo;
 
-import com.acuo.persist.entity.Portfolio;
 import com.acuo.persist.entity.Trade;
+import com.acuo.persist.entity.TradeValuation;
 import com.acuo.persist.entity.Valuation;
 import com.acuo.persist.ids.PortfolioId;
 import com.acuo.persist.services.PortfolioService;
@@ -18,25 +18,40 @@ import lombok.extern.slf4j.Slf4j;
 import javax.inject.Inject;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.TreeSet;
 
 @Slf4j
-public class PricingResultPersister implements ResultPersister<PricingResults> {
+public class PricingResultPersister implements ResultPersister<PricingResults>, MarkitValuationProcessor.PricingResultProcessor {
 
     private final TradeService<Trade> tradeService;
     private final ValuationService valuationService;
-    private final PortfolioService portfolioService;
     private final ValueService valueService;
+    private MarkitValuationProcessor.PricingResultProcessor nextProcessor;
 
     @Inject
-    public PricingResultPersister(TradeService<Trade> tradeService, ValuationService valuationService, PortfolioService portfolioService, ValueService valueService) {
+    public PricingResultPersister(TradeService<Trade> tradeService, ValuationService valuationService, ValueService valueService) {
         this.tradeService = tradeService;
         this.valuationService = valuationService;
-        this.portfolioService = portfolioService;
         this.valueService = valueService;
+    }
+
+    @Override
+    public MarkitValuationProcessor.ProcessorItem process(MarkitValuationProcessor.ProcessorItem processorItem) {
+        log.info("processing item {}", processorItem);
+        PricingResults results = processorItem.getResults();
+        Set<PortfolioId> portfolioIds = persist(results);
+        processorItem.setPortfolioIds(portfolioIds);
+        if (nextProcessor!= null)
+            return nextProcessor.process(processorItem);
+        else
+            return processorItem;
+    }
+
+    @Override
+    public void setNextProcessor(MarkitValuationProcessor.PricingResultProcessor nextProcessor) {
+        this.nextProcessor = nextProcessor;
     }
 
     public Set<PortfolioId> persist(PricingResults pricingResults) {
@@ -47,8 +62,7 @@ public class PricingResultPersister implements ResultPersister<PricingResults> {
             return portfolioIds;
         }
 
-        log.debug("persistMarkitResult start :" + pricingResults.getDate());
-
+        log.info("persisting {} markit result of {}", pricingResults.getResults().size(), pricingResults.getDate());
 
         LocalDate date = pricingResults.getDate();
         Currency currency = pricingResults.getCurrency();
@@ -58,14 +72,13 @@ public class PricingResultPersister implements ResultPersister<PricingResults> {
         for (Result<MarkitValuation> result : results) {
 
             for (Value value : result.getValue().getValues()) {
-
-
                 String tradeId = value.getTradeId();
-
-                log.debug("tradeId:" + tradeId);
-
                 Trade trade = tradeService.findById(tradeId);
 
+                if (trade == null) {
+                    log.warn("no trade with id {}", tradeId);
+                    continue;
+                }
 
                 Set<Valuation> valuations = trade.getValuations();
                 boolean found = false;
@@ -73,35 +86,25 @@ public class PricingResultPersister implements ResultPersister<PricingResults> {
                     for (Valuation valuation : valuations) {
                         if (valuation.getDate().equals(date)) {
                             //existing date, add or replace the value
-                            if (valuation.getValues() == null)
-                                valuation.setValues(new HashSet<com.acuo.persist.entity.Value>());
+
                             Set<com.acuo.persist.entity.Value> existedValues = valuation.getValues();
-
-
-                            for (com.acuo.persist.entity.Value existedValue : existedValues) {
-                                if (existedValue.getCurrency().equals(currency) && existedValue.getSource().equalsIgnoreCase("Markit")) {
-                                    log.debug("deleting value id [{}]", existedValue.getId());
-                                    try {
-                                        valueService.delete(existedValue.getId());
-                                        existedValues.remove(existedValue);
-                                    } catch (Exception e) {
+                            if (existedValues != null) {
+                                for (com.acuo.persist.entity.Value existedValue : existedValues) {
+                                    if (existedValue.getCurrency().equals(currency) && existedValue.getSource().equalsIgnoreCase("Markit")) {
+                                        log.debug("deleting value id [{}]", existedValue.getId());
+                                        try {
+                                            valueService.delete(existedValue.getId());
+                                            existedValues.remove(existedValue);
+                                        } catch (Exception e) {
+                                        }
                                     }
                                 }
                             }
 
-
-                            com.acuo.persist.entity.Value newValue = new com.acuo.persist.entity.Value();
-
-                            newValue.setSource("Markit");
-                            newValue.setCurrency(currency);
-                            newValue.setPv(value.getPv());
-
-                            valuation.getValues().add(newValue);
-
-                            valuationService.createOrUpdate(valuation);
-                            trade.getValuations().add(valuation);
+                            com.acuo.persist.entity.Value newValue = createValue(currency, value.getPv(), "Markit");
+                            newValue.setValuation(valuation);
+                            valueService.createOrUpdate(newValue);
                             portfolioIds.add(PortfolioId.fromString(trade.getPortfolio().getPortfolioId()));
-                            addsumValuationOfPortfolio(trade.getPortfolio(), date, currency, "Markit", value.getPv());
                             found = true;
                             break;
                         }
@@ -111,98 +114,35 @@ public class PricingResultPersister implements ResultPersister<PricingResults> {
                 if (!found) {
                     //new valutaion
 
-                    Valuation valuation = new Valuation();
+                    TradeValuation valuation = createValuation(date, trade.getTradeId());
+                    com.acuo.persist.entity.Value newValue = createValue(currency, value.getPv(), "Markit");
+                    newValue.setValuation(valuation);
 
-                    valuation.setValuationId(date.format(DateTimeFormatter.ofPattern("yyyy/MM/dd")) + "-" + trade.getTradeId());
-
-                    valuation.setDate(date);
-
-
-                    com.acuo.persist.entity.Value newValue = new com.acuo.persist.entity.Value();
-
-                    newValue.setSource("Markit");
-                    newValue.setCurrency(currency);
-                    newValue.setPv(value.getPv());
-
-                    Set<com.acuo.persist.entity.Value> values = new HashSet<com.acuo.persist.entity.Value>();
-
-                    values.add(newValue);
-
-                    valuation.setValues(values);
-
-                    if (trade.getValuations() != null)
-                        trade.getValuations().add(valuation);
-                    else {
-                        Set<Valuation> valuationSet = new HashSet<Valuation>();
-                        trade.setValuations(valuationSet);
-                    }
-
+                    valuation.setTrade(trade);
+                    valuation.setPortfolio(trade.getPortfolio());
 
                     valuationService.createOrUpdate(valuation);
-                    tradeService.createOrUpdate(trade);
                     portfolioIds.add(PortfolioId.fromString(trade.getPortfolio().getPortfolioId()));
-                    addsumValuationOfPortfolio(trade.getPortfolio(), date, currency, "Markit", value.getPv());
                 }
             }
         }
         return portfolioIds;
     }
 
-    private void addsumValuationOfPortfolio(Portfolio portfolio, LocalDate date, com.opengamma.strata.basics.currency.Currency currency, String source, Double pv) {
+    private TradeValuation createValuation(LocalDate date, String tradeId) {
+        TradeValuation valuation = new TradeValuation();
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy/MM/dd");
+        String dateFormatted = date.format(formatter);
+        valuation.setValuationId(dateFormatted + "-" + tradeId);
+        valuation.setDate(date);
+        return valuation;
+    }
 
-        portfolio = portfolioService.findById(portfolio.getPortfolioId(), 2);
-
-        Valuation theValuation = null;
-        com.acuo.persist.entity.Value theValue = null;
-
-
-        if (portfolio.getValuations() != null) {
-            for (Valuation valuation : portfolio.getValuations()) {
-                if (valuation.getDate().format(DateTimeFormatter.ofPattern("yyyy/MM/dd")).equals(date.format(DateTimeFormatter.ofPattern("yyyy/MM/dd")))) {
-                    theValuation = valuation;
-                    if (valuation.getValues() != null)
-                        for (com.acuo.persist.entity.Value value : valuation.getValues()) {
-                            if (value.getCurrency().equals(currency) && value.getSource().equals(source))
-                                theValue = value;
-                        }
-                }
-            }
-        }
-
-
-        if (theValuation == null) {
-            theValuation = new Valuation();
-            theValuation.setValuationId(date.format(DateTimeFormatter.ofPattern("yyyy/MM/dd")) + "-" + portfolio.getPortfolioId());
-            theValuation.setDate(date);
-            Set<com.acuo.persist.entity.Value> values = new HashSet<com.acuo.persist.entity.Value>();
-            theValue = new com.acuo.persist.entity.Value();
-            theValue.setPv(pv);
-            theValue.setCurrency(currency);
-            theValue.setSource(source);
-            values.add(theValue);
-            theValuation.setValues(values);
-
-            if (portfolio.getValuations() == null)
-                portfolio.setValuations(new HashSet<Valuation>());
-
-            portfolio.getValuations().add(theValuation);
-            portfolioService.createOrUpdate(portfolio);
-        } else {
-            if (theValue == null) {
-                theValue = new com.acuo.persist.entity.Value();
-                theValue.setPv(pv);
-                theValue.setCurrency(currency);
-                theValue.setSource(source);
-
-                if (theValuation.getValues() == null)
-                    theValuation.setValues(new HashSet<com.acuo.persist.entity.Value>());
-
-                theValuation.getValues().add(theValue);
-                valuationService.createOrUpdate(theValuation);
-            } else {
-                theValue.setPv(theValue.getPv() + pv);
-                valueService.createOrUpdate(theValue);
-            }
-        }
+    private com.acuo.persist.entity.Value createValue(Currency currency, Double pv, String source) {
+        com.acuo.persist.entity.Value newValue = new com.acuo.persist.entity.Value();
+        newValue.setSource(source);
+        newValue.setCurrency(currency);
+        newValue.setPv(pv);
+        return newValue;
     }
 }
